+++ bin/proving-service/src/commands/proxy.rs
@@ -115,15 +115,9 @@ impl StartProxy {
             info!("Metrics are not enabled");
         }
 
-        // Add status service
-        let status_service = ProxyStatusService::new(worker_lb);
-        let mut status_service = Service::new("status".to_string(), status_service);
-        status_service
-            .add_tcp(format!("{}:{}", PROXY_HOST, self.proxy_config.status_port).as_str());
-        info!(
-            "Status service listening on {}:{}/status",
-            PROXY_HOST, self.proxy_config.status_port
-        );
+        // Add gRPC status service
+        let status_service = ProxyStatusService::new(worker_lb, self.proxy_config.status_port);
+        info!("gRPC status service will start on port {}", self.proxy_config.status_port);
 
         server.add_service(health_check_service);
         server.add_service(update_workers_service);